evaluation/AndroidWorld/requirements.txt
absl-py==2.1.0
dm_env==1.6
fuzzywuzzy==0.18.0
google-generativeai==0.8.6  # 升级以支持 protobuf 6.x
google-ai-generativelanguage==0.6.15  # 必需依赖
grpcio==1.76.0  # 升级以支持 protobuf 6.x
grpcio-status==1.71.2
grpcio-tools==1.71.0  # 升级以支持 protobuf 6.x
immutabledict==2.0.0
IPython
jsonschema==4.17.3
matplotlib==3.6.1
numpy==1.26.3
# opencv-python==4.11.0.86  # 使用 conda 安装: conda install -c conda-forge opencv
pandas==2.1.4
pydub
protobuf>=6.0.0  # 必需：android_env 的 proto 文件是用 6.31.1 生成的，需要 6.x 版本
pysqlite3>=0.5.4  # 必需：用于 SQLite 操作（使用版本范围以支持不同平台）
python-Levenshtein
pytest
requests
tenacity
termcolor